弾の移動方向をﾗｼﾞｱﾝから度数法に変えたぜ

+++ Tohohu/PlayerShot.h
@@ -11,7 +11,7 @@ struct PlayerShot
 	Vector2 pos;			// À•W
 	Vector2 move;			// ˆÚ“®—Ê
 	float speed;			// ‘¬‚³
-	float moveAngle;		// “®‚­•ûŒü
+	int moveAngle;			// “®‚­•ûŒü
 	bool flag;				// •`‰æÌ×¸Ş
 	float endPos;			// •`‰æI—¹ˆÊ’u
 };
